Uploaded initial script file
Read-in routine will assume all rows (minus the header) of the .CSV are
valid commit lines and will write the content of that row to an AD
object. Script must back up all users first, which can either be done
per-line or all in one go at the beginning (not sure which is more
efficient).